new files and moving code around for formatting and readability
