+++ source/basic_locomotion_isaaclab/basic_locomotion_isaaclab/tasks/custom_events.py
@@ -119,44 +119,6 @@ def randomize_joint_parameters(
         )
 
 
-
-
-def randomize_joint_default_pos(
-        env: ManagerBasedEnv,
-        env_ids: torch.Tensor | None,
-        asset_cfg: SceneEntityCfg,
-        pos_distribution_params: tuple[float, float] | None = None,
-        operation: Literal["add", "scale", "abs"] = "abs",
-        distribution: Literal["uniform", "log_uniform", "gaussian"] = "uniform",
-):
-    """
-    Randomize the joint default positions which may be different from URDF due to calibration errors.
-    """
-    # extract the used quantities (to enable type-hinting)
-    asset: Articulation = env.scene[asset_cfg.name]
-
-    # resolve environment ids
-    if env_ids is None:
-        env_ids = torch.arange(env.scene.num_envs, device=asset.device)
-
-    # resolve joint indices
-    if asset_cfg.joint_ids == slice(None):
-        joint_ids = slice(None)  # for optimization purposes
-    else:
-        joint_ids = torch.tensor(asset_cfg.joint_ids, dtype=torch.int, device=asset.device)
-
-    if pos_distribution_params is not None:
-        pos = asset.data.default_joint_pos.to(asset.device).clone()
-        pos = _randomize_prop_by_op(
-            pos, pos_distribution_params, env_ids, joint_ids, operation=operation, distribution=distribution
-        )[env_ids][:, joint_ids]
-
-        if env_ids != slice(None) and joint_ids != slice(None):
-            env_ids = env_ids[:, None]
-        asset.data.default_joint_pos[env_ids, joint_ids] = pos
-
-
-
 def randomize_joint_friction_model(
     env: ManagerBasedEnv,
     env_ids: torch.Tensor | None,
@@ -260,25 +222,3 @@ def randomize_joint_delay_model(
                     second_order_delay_filter, second_order_delay_filter_distribution_params, env_ids, torch.arange(second_order_delay_filter.shape[1]), operation=operation, distribution=distribution
                 )[env_ids][:, actuator_joint_ids]
                 actuator.second_order_delay_filter[env_ids[:, None], actuator_joint_ids] = second_order_delay_filter
-
-
-def zero_command_velocity(
-    env: ManagerBasedEnv,
-    env_ids: torch.Tensor,
-):
-   
-    env._commands[env_ids, 0] = 0.0
-    env._commands[env_ids, 1] = 0.0
-    env._commands[env_ids, 2] = 0.0
-
-
-def resample_command_velocity(
-    env: ManagerBasedEnv,
-    env_ids: torch.Tensor,
-):
-   
-    # Sample new commands
-    env._commands[env_ids] = torch.zeros_like(env._commands[env_ids]).uniform_(-1.0, 1.0)
-    env._commands[env_ids, 0] *= 0.5 
-    env._commands[env_ids, 1] *= 0.25 
-    env._commands[env_ids, 2] *= 0.3
